Add InputDefaults to unify content padding and move to input package
