+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/molecule/input/InputDefaults.kt
@@ -1,0 +1,19 @@
+package app.k9mail.core.ui.compose.designsystem.molecule.input
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.Dp
+import app.k9mail.core.ui.compose.theme.MainTheme
+
+@Composable
+fun inputContentPadding(
+    start: Dp = MainTheme.spacings.double,
+    top: Dp = MainTheme.spacings.default,
+    end: Dp = MainTheme.spacings.double,
+    bottom: Dp = MainTheme.spacings.default,
+): PaddingValues = PaddingValues(
+    start = start,
+    top = top,
+    end = end,
+    bottom = bottom,
+)
